Basically rewrote everything
Because of how slow the earley parser got I decided to just do a hand written version of the compiler.

tldr; I removed everything and started a rewrite, and for couple hours of work I am quite proud of it :)

the readme is updated with info about the new iteration.

the lark file will be kept for now